% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{PlantCare_app}
\alias{PlantCare_app}
\title{Plant Care Shiny App}
\usage{
PlantCare_app(data = prepare_data())
}
\arguments{
\item{data}{Can be the 11 x 13 data frame obtained from final_data() or any
other object. Will be overwritten immediately with the data frame obtained
from the final_data() function.}
}
\value{
No return. There will be a pop-up with the user interface. Interrupt
R to stop the application (press "Stop" button or press Esc).
}
\description{
An interactive app that gives advice on whether to water, fertilize, repot,
or change light conditions for common houseplants. The advice is
plant-specific.Assumes the user is close
to their plant, i.e., to feel whether the soil is wet and see if the
leaves are dry or soft. Uses data scrapped from the website
https://plantsage.org/.
}
\examples{
# PlantCare_app()
}
