% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{prepare_data}
\alias{prepare_data}
\title{Scrap and Prepare Plant Data}
\usage{
prepare_data()
}
\arguments{
\item{...}{Does not contain arguments. The data is manually added as a
JavaScript object, cleaned up, restructured, and returned as a data frame.}
}
\value{
An 11 x 13 data frame with plant data for 11 common houseplants,
ready for use with the PlantCare app.
}
\description{
Scraps plant data from the website https://plantsage.org/.
Some new columns are added; these contain the same information as in the
data set https://plantsage.org/ but reformulated for easier use with the
PlantCare app. The reformulation was done by the package maintainer.
}
\examples{
data <- prepare_data()
View(data)
}
\references{
Plant Sage. (n.d.). Plant Sage. Retrieved May 31, 2024, from https://plantsage.org/
}
