% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{PlantCare_app}
\alias{PlantCare_app}
\title{Plant Care Shiny App}
\usage{
PlantCare_app(data = prepare_data())
}
\arguments{
\item{data}{Can be the 11 x 13 data frame obtained from final_data() or any
other object. Will be overwritten immediately with the data frame obtained
from the final_data() function.}
}
\value{
No return. There will be a pop-up with the user interface. Interrupt
R to stop the application (press "Stop" button or press Esc).
}
\description{
An interactive app that gives advice on how to care for common houseplants,
based on the plant species and its current state. Assumes the user is close
to their plant, i.e., to feel whether the soil is wet. Uses data scrapped
from https://plantsage.org/.
}
\examples{
# PlantCare_app()
}
\references{
Plant Sage. (n.d.). Plant Sage. Retrieved May 31, 2024,
from https://plantsage.org/ \cr
Plant.id. (n.d.). Plant.id. Retrieved May 31, 2024, from
https://plant.id/ \cr
Sweet, L., et al. (n.d.). How to repot a plant. wikiHow.
Retrieved May 31, 2024, from https://www.wikihow.com/Repot-a-Plant
}
